Updates +Connected user welcome page +Added user log page +Bug fixes +New route and functions active in the sidebar

+++ src/controllers/user_controller.ts
@@ -5,8 +5,10 @@ import role from "../models/role";
 import mongoose from "mongoose";
 import dotenv from "dotenv";
 import { verifyCreateUser, verifyQuestions } from "../middlewares/verifyForms";
+import { registerLog } from "../middlewares/verifySignup";
 import fs from "fs";
 import quest from "../models/quest";
+import { RequestUser } from "../types/types";
 dotenv.config();
 
 export const getUsers = async (req: Request, res: Response) => {
@@ -33,6 +35,7 @@ export const getUsers = async (req: Request, res: Response) => {
         limit:
             req.query && Number(req.query.limit) ? Number(req.query.limit) : 10,
         page: req.query && Number(req.query.page) ? Number(req.query.page) : 1,
+        sort:{created_at:-1},
         select: { password: 0 },
         populate: { path: "rol", select: { name: 1 } },
     };
@@ -206,7 +209,7 @@ export const createUser = async (req: any, res: Response) => {
 
         const savedUser = await newUser.save();
 
-        //await verifySignup.registerLog(req,`Registro al usuario: ${savedUser.firstname} ${savedUser.lastname} - cedula: ${savedUser.ci}`);
+        await registerLog(req,`Registro al usuario: ${savedUser.firstname} ${savedUser.lastname} - cedula: ${savedUser.ci}`);
 
         return res.json({ message: "Usuario Registrado", user: savedUser });
     } catch (err) {
@@ -244,17 +247,22 @@ export const validateUser = async (req: any, res: Response) => {
             password,
             position,
             questions,
-            answers,
+            answers
         } = req.body;
+
+        const avatar = req.file?.path || ''
 
         const checkRegister = await verifyCreateUser(
             req.body,
             true,
             req.userId
         );
+        const userFound = await user.findById(req.params.id)
+
+        if(!userFound) return res.status(404).json({message:'Usuario no encontrado'})
 
         const checkQuestions = await verifyQuestions(questions, answers);
-        
+
         if (checkQuestions !== "") {
             if(req.file && req.file.path)fs.unlinkSync(req.file.path || "");
             return res.status(400).json({ message: checkQuestions });
@@ -285,6 +293,12 @@ export const validateUser = async (req: any, res: Response) => {
             await newQuestion.save();
         }
 
+        if(avatar !== '' && userFound.avatar){
+            if (fs.existsSync(userFound.avatar)){
+                fs.unlinkSync(userFound.avatar || "")
+            }
+        }
+
         await user.updateOne(
             { _id: req.params.id },
             {
@@ -296,11 +310,13 @@ export const validateUser = async (req: any, res: Response) => {
                     password: await user.encryptPassword(password),
                     position,
                     first_login: false,
-                    avatar: req.file?.path || "",
+                    avatar: req.file?.path || userFound.avatar,
                     updated_at: new Date(),
                 },
             }
         );
+
+        await registerLog(req,`Se verifico usuario: ${userFound.firstname} ${userFound.lastname} - cedula: ${userFound.ci}`);
 
         return res.json({ message: "Usuario Verificdo" });
     } catch (err) {
@@ -335,9 +351,9 @@ export const updateUser = async (req: any, res: Response) => {
         if (checkRegister)
             return res.status(400).json({ message: checkRegister.message });
 
-        const foundUser = await user.findById(req.params.id || req.userId);
+        const userFound = await user.findById(req.params.id || req.userId);
 
-        if(!foundUser) return res.status(404).json({message:'Usuario no encontrado'})
+        if(!userFound) return res.status(404).json({message:'Usuario no encontrado'})
 
         const rolFind = await role.findOne({ name: { $in: req.body.rol } });
 
@@ -357,9 +373,9 @@ export const updateUser = async (req: any, res: Response) => {
             });
         }
 
-        if(avatar !== '' && foundUser.avatar){
-            if (fs.existsSync(foundUser.avatar)){
-                fs.unlinkSync(foundUser.avatar || "")
+        if(avatar !== '' && userFound.avatar){
+            if (fs.existsSync(userFound.avatar)){
+                fs.unlinkSync(userFound.avatar || "")
             }
         }
 
@@ -372,15 +388,15 @@ export const updateUser = async (req: any, res: Response) => {
                         firstname: req.body.firstname,
                         lastname: req.body.lastname,
                         email: req.body.email,
-                        password: req.body.allowPassword ? await user.encryptPassword(req.body.password) : foundUser.password,
+                        password: req.body.allowPassword ? await user.encryptPassword(req.body.password) : userFound.password,
                         rol: rolFind._id,
-                        avatar:avatar || foundUser.avatar,
+                        avatar:avatar || userFound.avatar,
                         position:req.body.position
                     },
                 }
             );
        
-        //await verifySignup.registerLog(req,`Modifico usuario: ${foundUser.firstname} ${foundUser.lastname} - cedula: ${foundUser.ci}`);
+        await registerLog(req,`Modifico usuario: ${userFound.firstname} ${userFound.lastname} - cedula: ${userFound.ci}`);
 
         return res.json({ message: "Usuario modificado" });
     } catch (err) {
@@ -392,7 +408,7 @@ export const updateUser = async (req: any, res: Response) => {
     }
 };
 
-export const deleteUser = async (req: Request, res: Response) => {
+export const deleteUser = async (req: RequestUser, res: Response) => {
     try {
         const userFind = await user.findById(req.params.id);
         const listUsers = await user.paginate({}, {});
@@ -412,6 +428,8 @@ export const deleteUser = async (req: Request, res: Response) => {
 
         await quest.deleteMany({user:req.params.id})
         await user.findByIdAndDelete(req.params.id);
+
+        await registerLog(req,`Elimino usuario: ${userFind.firstname} ${userFind.lastname} - cedula: ${userFind.ci}`);
 
         return res.json({ message: "Usuario eliminado" });
     } catch (err) {
